+++ chatroom_gp2/src/main/java/com/example/chatroom_gp2/controllers/UserController.java
@@ -10,6 +10,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("users")
@@ -24,9 +25,13 @@ public class UserController {
     }
 
     @GetMapping(value = "/{id}")
-        public ResponseEntity<User> getUserByID(@PathVariable long id){
-            return new ResponseEntity<>(userService.getUserById(id), HttpStatus.OK);
+    public ResponseEntity<User> getUserByID(@PathVariable long id){
+        Optional<User> user = userService.getUserById(id);
+        if(user.isPresent()){
+            return new ResponseEntity<>(user.get(), HttpStatus.OK);
         }
+        return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+    }
 
     @PostMapping
     public ResponseEntity<User> createUser(@RequestBody User user){
@@ -36,13 +41,21 @@ public class UserController {
 
     @PatchMapping(value = "/{id}")
     public ResponseEntity<User> updateUser(@RequestBody UserDTO userDTO, @PathVariable long id){
-        return new ResponseEntity<>(userService.updateUser(userDTO, id), HttpStatus.OK);
+        Optional<User> user = userService.getUserById(id);
+        if(user.isPresent()) {
+            return new ResponseEntity<>(userService.updateUser(userDTO, id), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<Long> deleteUserById(@PathVariable long id){
-        userService.deleteUserById(id);
-        return  new ResponseEntity<>(id, HttpStatus.OK );
+        Optional<User> user = userService.getUserById(id);
+        if(user.isPresent()) {
+            userService.deleteUserById(id);
+            return new ResponseEntity<>(id, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
     }
 
 }
